+++ src/App/Hooks/ScriptLoader.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace WpTheme\Scaffold\Core\Hooks;
+namespace WpTheme\Scaffold\App\Hooks;
 
 use WpTheme\Scaffold\Core\Abstracts\Hooks;
 
-use WpTheme\Scaffold\Core\Services\HookLoader;
+use WpTheme\Scaffold\App\Services\HookLoader;
 
 final class ScriptLoader extends Hooks {
