+++ src/App/Services/HookLoader.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace WpTheme\Scaffold\Core\Services;
+namespace WpTheme\Scaffold\App\Services;
 
 use WpTheme\Scaffold\Core\Resources\Hook;
